Fixed small memory leak in Aapt
This change updates the way Bundle handles the ManifestMinSDKVersion
setting. This is changed from a char* to a String8. This is done
to keep in line with the style of Bundle, that is to contain in String8
instances strings and settings that are local to it and only maintain
char* members which point to memory that is managed by other classes/functions.
Since this particular char* was not managed by another class this resulted in
a small memory leak. The public API for Bundle remains unchanged.
The code that sets this paramater of bundle has also been modified to get
rid of the offending call to strdup which caused the leak and is no longer
necessary.

Change-Id: Ic567ea9d22ce5eadf4c36ef6dd2d3c34a6eeb8c0