Fix leak of global references and AudioFocusDeathHandler objects.
When an AudioService is required, AudioFocusDeathHandler object was created,
but it never deleted. This causes the system crash after the number of
registered objects into JNI Global Table exceeded the torelance.
In the patch, unlinkToDeath() is called to unlink AudioFocusDeathHandler
object when it was removed from the stack, and the object be the
garbage collected.

This fix was already a part of ICS. Back-ported for GB.
Ref: OMAPS00260211

Change-Id: Icdc05604d6262a8c82037fda6ea6e8762bcc6f22
Signed-off-by: Nanako BABA <baba@ti.com>

Conflicts:

	media/java/android/media/AudioService.java

resolved